+++ ChangeLog.md
@@ -7,6 +7,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 
 
+## [UNRELEASED]
+
+### fixed
+
+- fixed Java Interop callbacks to support import statements
+
+
+
 ## [0.7.0] - 2018-07-31
 
 ### Added